Renamed Node::Type to Node::Kind
The thinking here was to allow distinction between types declared in
Iron code and the enumeration that lists the various types of nodes
in the Iron AST. Because it's short and distinct, Iron nodes now
have a Kind enum. The actual structures in the Iron code all have
Types, which are themselves Iron AST Nodes that each have a Kind.